Split the structures into primitives and pml + more edits.
