+++ beamz/design/__init__.py
@@ -3,10 +3,11 @@ Design module for BEAMZ - Contains components for designing photonic structures.
 """
 
 from beamz.design.materials import Material, CustomMaterial
-from beamz.design.structures import (
-    Design, Rectangle, Circle, Ring,
-    CircularBend, Polygon, Taper
+from beamz.design.structures import Design
+from beamz.design.primitives import (
+    Rectangle, Circle, Ring, CircularBend, Polygon, Taper
 )
+from beamz.design.pml import PML
 from beamz.design.sources import ModeSource, GaussianSource
 from beamz.design.monitors import Monitor
 from beamz.design.signals import ramped_cosine, plot_signal
@@ -15,7 +16,7 @@ from beamz.design.mode import solve_modes, slab_mode_source
 __all__ = [
     'Material', 'CustomMaterial',
     'Design', 'Rectangle', 'Circle', 'Ring',
-    'CircularBend', 'Polygon', 'Taper',
+    'CircularBend', 'Polygon', 'Taper', 'PML',
     'ModeSource', 'GaussianSource',
     'Monitor',
     'ramped_cosine', 'plot_signal',
