+++ beamz/design/structures.py
@@ -1,16 +1,16 @@
-import matplotlib.pyplot as plt
-from matplotlib.patches import Rectangle as MatplotlibRectangle, PathPatch, Circle as MatplotlibCircle
-from matplotlib.path import Path
 import random
 import numpy as np
-from beamz.design.materials import Material
+import gdspy
+
 from beamz import viz as viz
-from beamz.const import µm, EPS_0, MU_0
+from beamz.const import µm
+from beamz.helpers import display_status
+
+from beamz.design.materials import Material
 from beamz.design.sources import ModeSource, GaussianSource
 from beamz.design.monitors import Monitor
-from beamz.helpers import display_status
-import colorsys
-import gdspy
+from beamz.design.primitives import Polygon, Rectangle, Circle, Ring, CircularBend, Taper
+from beamz.design.pml import PML
 
 class Design:
     def __init__(self, width=5*µm, height=5*µm, depth=0, material=None, color=None, border_color="black", 
@@ -605,663 +605,3 @@ class Design:
         new_design.layers = self.layers.copy() if hasattr(self, 'layers') else {}
         
         return new_design
-
-
-
-class Polygon:
-    def __init__(self, vertices=None, material=None, color=None, optimize=False, interiors=None, depth=0, z=0):
-        # First ensure vertices are properly formatted and ordered
-        self.vertices = self._process_vertices(vertices if vertices is not None else [], z)
-        # Preserve interior path orientation (needed for proper hole rendering with nonzero rule)
-        self.interiors = [self._process_vertices_preserve_orientation(interior, z)
-                            for interior in (interiors if interiors is not None else [])]
-        self.material = material
-        self.optimize = optimize
-        self.color = color if color is not None else self.get_random_color_consistent()
-        self.depth = depth if depth is not None else 0
-        self.z = z if z is not None else 0
-    
-    def _process_vertices(self, vertices, z=0):
-        """Process vertices to ensure they are 3D and properly ordered counterclockwise."""
-        if not vertices: return []   
-        # First ensure all vertices are 3D
-        vertices_3d = self._ensure_3d_vertices(vertices)
-        # Project to 2D for CCW check
-        vertices_2d = [(v[0], v[1]) for v in vertices_3d]
-        # Ensure counterclockwise ordering
-        if len(vertices_2d) >= 3:
-            vertices_2d = self._ensure_ccw_vertices(vertices_2d)
-            # Convert back to 3D maintaining original z-coordinates or using provided z
-            vertices_3d = [(x, y, vertices_3d[i][2] if len(vertices_3d[i]) > 2 else z) 
-                          for i, (x, y) in enumerate(vertices_2d)]
-        return vertices_3d
-
-    def _process_vertices_preserve_orientation(self, vertices, z=0):
-        """Ensure 3D coordinates but preserve original vertex order (for holes)."""
-        if not vertices: return []
-        vertices_3d = self._ensure_3d_vertices(vertices)
-        # Do not alter orientation; only ensure z coordinate presence
-        return [(v[0], v[1], v[2] if len(v) > 2 else z) for v in vertices_3d]
-    
-    def _ensure_ccw_vertices(self, vertices_2d):
-        """Ensure vertices are ordered counterclockwise by computing signed area."""
-        if len(vertices_2d) < 3: return vertices_2d
-        # Calculate signed area
-        area = 0
-        for i in range(len(vertices_2d)):
-            j = (i + 1) % len(vertices_2d)
-            area += vertices_2d[i][0] * vertices_2d[j][1]
-            area -= vertices_2d[j][0] * vertices_2d[i][1]
-        # If area is positive, vertices are already CCW, else: reverse the vertices
-        if area < 0: return vertices_2d[::-1]
-        return vertices_2d
-    
-    def _ensure_3d_vertices(self, vertices):
-        """Convert 2D vertices (x,y) to 3D vertices (x,y,z) with z=0 if not provided."""
-        if not vertices: return []
-        result = []
-        for v in vertices:
-            if len(v) == 2: result.append((v[0], v[1], 0.0))  # Add z=0 for 2D vertices
-            elif len(v) == 3: result.append(v)  # Keep 3D vertices as-is
-            else: raise ValueError(f"Vertex must have 2 or 3 coordinates, got {len(v)}")
-        return result
-    
-    def _vertices_2d(self, vertices=None):
-        """Get 2D projection of vertices for plotting (x,y only)."""
-        if vertices is None: vertices = self.vertices
-        return [(v[0], v[1]) for v in vertices]
-    
-    def get_random_color_consistent(self, saturation=0.6, value=0.7):
-        """Generate a random color with consistent perceived brightness and saturation."""
-        hue = random.random() # Generate random hue (0-1)
-        r, g, b = colorsys.hsv_to_rgb(hue, saturation, value)
-        return '#{:02x}{:02x}{:02x}'.format(int(r * 255), int(g * 255), int(b * 255))
-    
-    def shift(self, x, y, z=0):
-        """Shift the polygon by (x,y,z) and return self for method chaining. z is optional for 2D compatibility."""
-        if self.vertices: self.vertices = [(v[0] + x, v[1] + y, v[2] + z) for v in self.vertices]
-        new_interiors_paths = []
-        for interior_path in self.interiors:
-            if interior_path: new_interiors_paths.append([(v[0] + x, v[1] + y, v[2] + z) for v in interior_path])
-        self.interiors = new_interiors_paths
-        return self
-    
-    def scale(self, s_x, s_y=None, s_z=None):
-        """Scale the polygon around its center of mass and return self for method chaining.
-        If s_y and s_z are None, uniform scaling is applied to all dimensions."""
-        if s_y is None: s_y = s_x  # Uniform scaling in xy if only one parameter given
-        if s_z is None: s_z = 1.0 if s_y != s_x else s_x  # Keep z unchanged for 2D, or uniform for 3D
-        
-        if self.vertices:
-            # Calculate center of mass of the exterior
-            x_center = sum(v[0] for v in self.vertices) / len(self.vertices)
-            y_center = sum(v[1] for v in self.vertices) / len(self.vertices)
-            z_center = sum(v[2] for v in self.vertices) / len(self.vertices)
-            # Scale exterior
-            self.vertices = [(x_center + (v[0] - x_center) * s_x,
-                              y_center + (v[1] - y_center) * s_y,
-                              z_center + (v[2] - z_center) * s_z)
-                              for v in self.vertices]
-            # Scale interiors
-            new_interiors_paths = []
-            for interior_path in self.interiors:
-                if interior_path:
-                    new_interiors_paths.append([(x_center + (v[0] - x_center) * s_x,
-                                                 y_center + (v[1] - y_center) * s_y,
-                                                 z_center + (v[2] - z_center) * s_z)
-                                                 for v in interior_path])
-            self.interiors = new_interiors_paths
-        return self
-    
-    def rotate(self, angle, axis='z', point=None):
-        """Rotate the polygon around its center of mass or specified point.
-        angle: Rotation angle in degrees
-        axis: Rotation axis ('x', 'y', or 'z'). Default 'z' for 2D compatibility
-        point: Optional (x,y,z) point to rotate around. If None, rotates around center of exterior.
-        """
-        if self.vertices:
-            angle_rad = np.radians(angle)
-            if point is None:
-                # Calculate center of mass of the exterior
-                x_center = sum(v[0] for v in self.vertices) / len(self.vertices)
-                y_center = sum(v[1] for v in self.vertices) / len(self.vertices)
-                z_center = sum(v[2] for v in self.vertices) / len(self.vertices)
-            else: x_center, y_center, z_center = (point[0], point[1], point[2] if len(point) > 2 else 0)
-
-            # Define rotation matrices for each axis
-            if axis == 'z':  # 2D rotation in xy plane (most common)
-                cos_a, sin_a = np.cos(angle_rad), np.sin(angle_rad)
-                self.vertices = [
-                    (x_center + (v[0] - x_center) * cos_a - (v[1] - y_center) * sin_a,
-                     y_center + (v[0] - x_center) * sin_a + (v[1] - y_center) * cos_a,
-                     v[2])
-                    for v in self.vertices
-                ]
-            elif axis == 'x':  # Rotation around x-axis (yz plane)
-                cos_a, sin_a = np.cos(angle_rad), np.sin(angle_rad)
-                self.vertices = [
-                    (v[0],
-                     y_center + (v[1] - y_center) * cos_a - (v[2] - z_center) * sin_a,
-                     z_center + (v[1] - y_center) * sin_a + (v[2] - z_center) * cos_a)
-                    for v in self.vertices
-                ]
-            elif axis == 'y':  # Rotation around y-axis (xz plane)
-                cos_a, sin_a = np.cos(angle_rad), np.sin(angle_rad)
-                self.vertices = [
-                    (x_center + (v[0] - x_center) * cos_a + (v[2] - z_center) * sin_a,
-                     v[1],
-                     z_center - (v[0] - x_center) * sin_a + (v[2] - z_center) * cos_a)
-                    for v in self.vertices
-                ]
-            else:
-                raise ValueError(f"Invalid rotation axis '{axis}'. Must be 'x', 'y', or 'z'.")
-            
-            # Rotate interiors
-            new_interiors_paths = []
-            for interior_path in self.interiors:
-                if interior_path:
-                    if axis == 'z':
-                        new_interiors_paths.append([
-                            (x_center + (v[0] - x_center) * cos_a - (v[1] - y_center) * sin_a,
-                             y_center + (v[0] - x_center) * sin_a + (v[1] - y_center) * cos_a,
-                             v[2])
-                            for v in interior_path
-                        ])
-                    elif axis == 'x':
-                        new_interiors_paths.append([
-                            (v[0],
-                             y_center + (v[1] - y_center) * cos_a - (v[2] - z_center) * sin_a,
-                             z_center + (v[1] - y_center) * sin_a + (v[2] - z_center) * cos_a)
-                            for v in interior_path
-                        ])
-                    elif axis == 'y':
-                        new_interiors_paths.append([
-                            (x_center + (v[0] - x_center) * cos_a + (v[2] - z_center) * sin_a,
-                             v[1],
-                             z_center - (v[0] - x_center) * sin_a + (v[2] - z_center) * cos_a)
-                            for v in interior_path
-                        ])
-            self.interiors = new_interiors_paths
-        return self
-
-    def add_to_plot(self, ax, facecolor=None, edgecolor="black", alpha=None, linestyle=None):
-        """Delegate polygon drawing to beamz.viz.draw_polygon."""
-        return viz.draw_polygon(ax, self, facecolor=facecolor, edgecolor=edgecolor, alpha=alpha, linestyle=linestyle)
-
-    def copy(self):
-        # Ensure interiors are copied as lists of tuples/lists, not as references
-        copied_interiors = [list(path) for path in self.interiors if path] if self.interiors else []
-        return Polygon(vertices=list(self.vertices) if self.vertices else [], 
-                       interiors=copied_interiors, 
-                       material=self.material, # Material can be shared
-                       color=self.color, 
-                       optimize=self.optimize,
-                       depth=self.depth,
-                       z=self.z)
-        
-    def get_bounding_box(self):
-        """Get the bounding box of the polygon as (min_x, min_y, min_z, max_x, max_y, max_z)"""
-        if not self.vertices or len(self.vertices) == 0:
-            return (0, 0, 0, 0, 0, 0)
-        # Extract x, y, and z coordinates
-        x_coords = [v[0] for v in self.vertices]
-        y_coords = [v[1] for v in self.vertices]
-        z_coords = [v[2] for v in self.vertices]
-        # Calculate min and max
-        min_x, max_x = min(x_coords), max(x_coords)
-        min_y, max_y = min(y_coords), max(y_coords)
-        min_z, max_z = min(z_coords), max(z_coords)
-        return (min_x, min_y, min_z, max_x, max_y, max_z)
-        
-    def _point_in_polygon_single_path(self, x, y, path_vertices):
-        """Check if a point is inside a single, simple polygon path using ray-casting.
-        Uses 2D projection (xy plane) for 3D vertices."""
-        if not path_vertices: return False
-        # Project to 2D for point-in-polygon test
-        path_2d = self._vertices_2d(path_vertices)
-        n = len(path_2d)
-        inside = False
-        p1x, p1y = path_2d[0]
-        for i in range(n + 1):
-            p2x, p2y = path_2d[i % n] # Ensure closure for ray casting
-            if y > min(p1y, p2y):
-                if y <= max(p1y, p2y):
-                    if x <= max(p1x, p2x):
-                        if p1y != p2y: # Avoid division by zero for horizontal lines
-                            xinters = (y - p1y) * (p2x - p1x) / (p2y - p1y) + p1x
-                        else: # Edge is horizontal
-                            xinters = p1x # Doesn't matter, will compare x <= max(p1x, p2x)
-                        if p1x == p2x or x <= xinters: # For vertical edge or point left of intersection
-                            inside = not inside
-            p1x, p1y = p2x, p2y
-        return inside
-
-    def point_in_polygon(self, x, y, z=None):
-        """Check if a point (x,y,z) is inside this polygon (which may have holes).
-        For 3D, currently uses 2D projection in xy plane. z parameter is for future 3D containment."""
-        # Use self.vertices for exterior and self.interiors for holes
-        exterior_path = self.vertices
-        interior_paths = self.interiors
-        if not exterior_path: return False
-        # Check if point is in the exterior boundary (2D projection)
-        if not self._point_in_polygon_single_path(x, y, exterior_path):
-            return False # Not in exterior, so definitely not in polygon
-        # If in exterior, check if it's in any of the holes
-        for interior_path_pts in interior_paths:
-            if interior_path_pts and self._point_in_polygon_single_path(x, y, interior_path_pts):
-                return False # Point is in a hole, so not in polygon
-        return True # In exterior and not in any hole
-
-class Rectangle(Polygon):
-    def __init__(self, position=(0,0,0), width=1, height=1, depth=1, material=None, color=None, is_pml=False, 
-                    optimize=False, z=None):
-        # Handle z parameter for backward compatibility
-        if z is not None:
-            if len(position) == 2: position = (position[0], position[1], z)
-            elif len(position) == 3: position = (position[0], position[1], z)  # Override z from position
-        # Handle 2D position input (x,y) by adding z=0
-        if len(position) == 2: position = (position[0], position[1], 0.0)
-        elif len(position) == 3: position = position
-        else: raise ValueError("Position must be (x,y) or (x,y,z)")
-        # Calculate vertices for the rectangle in 3D
-        x, y, z_pos = position
-        vertices = [(x, y, z_pos),  # Bottom left
-                    (x + width, y, z_pos),  # Bottom right
-                    (x + width, y + height, z_pos),  # Top right
-                    (x, y + height, z_pos)]
-        super().__init__(vertices=vertices, material=material, color=color, optimize=optimize, depth=depth, z=z_pos)
-        self.position = position
-        self.width = width
-        self.height = height
-        self.depth = depth
-        self.is_pml = is_pml
-        
-    def get_bounding_box(self):
-        """Get the axis-aligned bounding box for this rectangle."""
-        # For non-rotated rectangles, this is straightforward
-        if not hasattr(self, 'vertices') or len(self.vertices) == 0:
-            x, y, z = self.position
-            return (x, y, z, x + self.width, y + self.height, z + self.depth)
-        # For potentially rotated rectangles, use the vertices
-        return super().get_bounding_box()
-
-    def shift(self, x, y, z=0):
-        """Shift the rectangle by (x,y,z) and return self for method chaining."""
-        self.position = (self.position[0] + x, self.position[1] + y, self.position[2] + z)
-        super().shift(x, y, z)
-        return self
-
-    def rotate(self, angle, axis='z', point=None):
-        """Rotate the rectangle around its center of mass or specified point."""        
-        # Use parent class rotation method
-        super().rotate(angle, axis, point)
-        # Calculate new bounding box after rotation
-        min_x = min(v[0] for v in self.vertices)
-        min_y = min(v[1] for v in self.vertices)
-        min_z = min(v[2] for v in self.vertices)
-        max_x = max(v[0] for v in self.vertices)
-        max_y = max(v[1] for v in self.vertices)
-        max_z = max(v[2] for v in self.vertices)
-        # Update position to be the bottom-left corner
-        self.position = (min_x, min_y, min_z)
-        self.width = max_x - min_x
-        self.height = max_y - min_y
-        self.depth = max_z - min_z
-        return self
-
-    def scale(self, s_x, s_y=None, s_z=None):
-        """Scale the rectangle around its center of mass and return self for method chaining."""
-        if s_y is None: s_y = s_x  # Uniform scaling in xy if only one parameter given
-        if s_z is None: s_z = 1.0 if s_y != s_x else s_x  # Keep z unchanged for 2D, or uniform for 3D
-        super().scale(s_x, s_y, s_z)
-        self.width *= s_x
-        self.height *= s_y
-        self.depth *= s_z
-        return self
-    
-    def copy(self):
-        """Create a copy of this rectangle with the same attributes and vertices."""
-        new_rect = Rectangle(self.position, self.width, self.height, self.depth, 
-                            self.material, self.color, self.is_pml, self.optimize)
-        # Ensure vertices are copied exactly as they are (important for rotated rectangles)
-        new_rect.vertices = [(x, y, z) for x, y, z in self.vertices]
-        return new_rect
-
-class Circle(Polygon):
-    def __init__(self, position=(0,0), radius=1, points=32, material=None, color=None, optimize=False, depth=0, z=0):
-        # Handle 2D position input (x,y) by adding z=0
-        if len(position) == 2: position = (position[0], position[1], 0.0)
-        elif len(position) == 3: position = position
-        else: raise ValueError("Position must be (x,y) or (x,y,z)")
-        theta = np.linspace(0, 2*np.pi, points, endpoint=False)
-        vertices = [(position[0] + radius * np.cos(t), position[1] + radius * np.sin(t), position[2]) for t in theta]
-        super().__init__(vertices=vertices, material=material, color=color, optimize=optimize, depth=depth, z=z)
-        self.position = position
-        self.radius = radius
-        self.points = points
-    
-    def shift(self, x, y, z=0):
-        """Shift the circle by (x,y,z) and return self for method chaining."""
-        self.position = (self.position[0] + x, self.position[1] + y, self.position[2] + z)
-        super().shift(x, y, z)
-        return self
-    
-    def scale(self, s_x, s_y=None, s_z=None):
-        """Scale the circle radius and return self for method chaining.
-        For circles, s_x is used as the radius scaling factor. s_y and s_z affect shape if different."""
-        if s_y is None: s_y = s_x  # Uniform scaling in xy if only one parameter given
-        if s_z is None: s_z = 1.0   # Don't scale z for circles by default
-        self.radius *= s_x
-        # Regenerate vertices with new radius
-        theta = np.linspace(0, 2*np.pi, self.points, endpoint=False)
-        self.vertices = [(self.position[0] + self.radius * np.cos(t), 
-                         self.position[1] + self.radius * np.sin(t),
-                         self.position[2]) for t in theta]
-        return self
-    
-    def copy(self):
-        return Circle(position=self.position, radius=self.radius, points=self.points, 
-                     material=self.material, color=self.color, optimize=self.optimize, 
-                     depth=self.depth, z=self.z)
-
-class Ring(Polygon):
-    def __init__(self, position=(0,0), inner_radius=1, outer_radius=2, material=None, color=None,
-                    optimize=False, points=256, depth=0, z=None):
-        # Handle z parameter for backward compatibility
-        if z is not None:
-            if len(position) == 2: position = (position[0], position[1], z)
-            elif len(position) == 3: position = (position[0], position[1], z)  # Override z from position
-        
-        # Handle 2D position input (x,y) by adding z=0
-        if len(position) == 2: position = (position[0], position[1], 0.0)
-        elif len(position) == 3: position = position
-        else: raise ValueError("Position must be (x,y) or (x,y,z)")
-        theta = np.linspace(0, 2*np.pi, points, endpoint=False) # CCW, N points
-        # Exterior path (CCW, N points)
-        # These are unclosed paths, as expected by Polygon.add_to_plot's Path logic
-        outer_ext_vertices = [(position[0] + outer_radius * np.cos(t), 
-                               position[1] + outer_radius * np.sin(t),
-                               position[2]) for t in theta]
-        # Interior path (should be CW for Matplotlib Path hole convention if exterior is CCW)
-        # Generate points CW by reversing theta or using reversed(theta)
-        inner_int_vertices_cw = [(position[0] + inner_radius * np.cos(t), 
-                                  position[1] + inner_radius * np.sin(t),
-                                  position[2]) for t in reversed(theta)]
-        super().__init__(vertices=outer_ext_vertices, 
-                         interiors=[inner_int_vertices_cw] if inner_int_vertices_cw else [], 
-                         material=material, color=color, optimize=optimize, depth=depth, z=position[2])
-        self.points = points # Store for potential regeneration or other logic if needed
-        self.position = position
-        self.inner_radius = inner_radius
-        self.outer_radius = outer_radius
-    
-    def shift(self, x, y, z=0):
-        """Shift the ring by (x,y,z) and return self for method chaining."""
-        self.position = (self.position[0] + x, self.position[1] + y, self.position[2] + z)
-        super().shift(x, y, z)
-        return self
-    
-    def scale(self, s_x, s_y=None, s_z=None):
-        """Scale the ring radii and return self for method chaining."""
-        if s_y is None: s_y = s_x  # Uniform scaling in xy if only one parameter given
-        if s_z is None: s_z = 1.0   # Don't scale z for rings by default
-        self.inner_radius *= s_x; self.outer_radius *= s_x
-        # Regenerate vertices with new radii
-        theta = np.linspace(0, 2*np.pi, self.points, endpoint=False)
-        # Outer circle points (exterior) and inner circle points (interior hole)
-        outer_vertices = [(self.position[0] + self.outer_radius * np.cos(t), 
-                          self.position[1] + self.outer_radius * np.sin(t),
-                          self.position[2]) for t in theta]
-        inner_vertices = [(self.position[0] + self.inner_radius * np.cos(t), 
-                          self.position[1] + self.inner_radius * np.sin(t),
-                          self.position[2]) for t in reversed(theta)]
-        self.vertices = outer_vertices
-        self.interiors = [inner_vertices]
-        return self
-    
-    def add_to_plot(self, ax, facecolor=None, edgecolor="black", alpha=None, linestyle=None):
-        if facecolor is None: facecolor = self.color
-        if alpha is None: alpha = 1
-        if linestyle is None: linestyle = '-'
-        # Use the generic Polygon.add_to_plot which now handles holes via self.vertices and self.interiors
-        # Ring.__init__ now populates self.vertices (exterior) and self.interiors correctly.
-        super().add_to_plot(ax, facecolor=facecolor, edgecolor=edgecolor, alpha=alpha, linestyle=linestyle)
-
-    def copy(self):
-        # Ring's __init__ now correctly sets up vertices and interiors for Polygon base class
-        return Ring(position=self.position, 
-                    inner_radius=self.inner_radius, 
-                    outer_radius=self.outer_radius, 
-                    material=self.material, # Material can be shared
-                    color=self.color, 
-                    optimize=self.optimize,
-                    points=self.points,
-                    depth=self.depth,
-                    z=self.z)
-
-class CircularBend(Polygon):
-    def __init__(self, position=(0,0), inner_radius=1, outer_radius=2, angle=90, rotation=0, material=None, 
-                 facecolor=None, optimize=False, points=64, depth=0, z=0):
-        # Handle 2D position input (x,y) by adding z=0
-        if len(position) == 2: position = (position[0], position[1], 0.0)
-        elif len(position) == 3: position = position
-        else: raise ValueError("Position must be (x,y) or (x,y,z)")
-        self.points = points
-        theta = np.linspace(0, np.radians(angle), points)
-        rotation_rad = np.radians(rotation)
-        outer_vertices = [(position[0] + outer_radius * np.cos(t + rotation_rad),
-                          position[1] + outer_radius * np.sin(t + rotation_rad),
-                          position[2]) for t in theta]
-        inner_vertices = [(position[0] + inner_radius * np.cos(t + rotation_rad),
-                          position[1] + inner_radius * np.sin(t + rotation_rad),
-                          position[2]) for t in reversed(theta)]
-        vertices = outer_vertices + inner_vertices
-        super().__init__(vertices=vertices, material=material, color=facecolor, optimize=optimize, depth=depth, z=z)
-        self.position = position
-        self.inner_radius = inner_radius
-        self.outer_radius = outer_radius
-        self.angle = angle
-        self.rotation = rotation
-    
-    def shift(self, x, y, z=0):
-        """Shift the bend by (x,y,z) and return self for method chaining."""
-        self.position = (self.position[0] + x, self.position[1] + y, self.position[2] + z)
-        super().shift(x, y, z)
-        return self
-    
-    def rotate(self, angle, axis='z', point=None):
-        """Rotate the bend around its center or specified point."""
-        if axis == 'z':  # Only z-rotation makes sense for circular bends in xy plane
-            self.rotation = (self.rotation + angle) % 360
-        super().rotate(angle, axis, point or self.position)
-        return self
-    
-    def scale(self, s_x, s_y=None, s_z=None):
-        """Scale the bend radii and return self for method chaining."""
-        if s_y is None: s_y = s_x  # Uniform scaling in xy if only one parameter given
-        if s_z is None: s_z = 1.0   # Don't scale z for bends by default
-        
-        self.inner_radius *= s_x; self.outer_radius *= s_x
-        theta = np.linspace(0, np.radians(self.angle), self.points)
-        rotation_rad = np.radians(self.rotation)
-        outer_vertices = [(self.position[0] + self.outer_radius * np.cos(t + rotation_rad),
-                          self.position[1] + self.outer_radius * np.sin(t + rotation_rad),
-                          self.position[2]) for t in theta]
-        inner_vertices = [(self.position[0] + self.inner_radius * np.cos(t + rotation_rad),
-                          self.position[1] + self.inner_radius * np.sin(t + rotation_rad),
-                          self.position[2]) for t in reversed(theta)]
-        self.vertices = outer_vertices + inner_vertices
-        return self
-    
-    def add_to_plot(self, ax, facecolor=None, edgecolor="black", alpha=None, linestyle=None):
-        if facecolor is None: facecolor = self.color
-        if alpha is None: alpha = 1
-        if linestyle is None: linestyle = '-'
-        # Convert angles to radians
-        angle_rad = np.radians(self.angle)
-        rotation_rad = np.radians(self.rotation)
-        theta = np.linspace(rotation_rad, rotation_rad + angle_rad, self.points, endpoint=True)
-        # Outer and inner arc points (project 3D to 2D for plotting)
-        x_outer = self.position[0] + self.outer_radius * np.cos(theta)
-        y_outer = self.position[1] + self.outer_radius * np.sin(theta)
-        x_inner = self.position[0] + self.inner_radius * np.cos(theta)
-        y_inner = self.position[1] + self.inner_radius * np.sin(theta)
-        # Create a closed path by combining points and adding connecting lines
-        vertices = np.vstack([
-            [x_outer[0], y_outer[0]],
-            *np.column_stack([x_outer[1:], y_outer[1:]]),
-            [x_inner[-1], y_inner[-1]],
-            *np.column_stack([x_inner[-2::-1], y_inner[-2::-1]]),
-            [x_outer[0], y_outer[0]]
-        ])
-        # Define path codes for a single continuous path
-        codes = [Path.MOVETO] + \
-                [Path.LINETO] * (len(vertices) - 2) + \
-                [Path.CLOSEPOLY]
-        # Create the path and patch
-        path = Path(vertices, codes)
-        bend_patch = PathPatch(path, facecolor=facecolor, alpha=alpha, edgecolor=edgecolor, linestyle=linestyle)
-        ax.add_patch(bend_patch)
-        
-    def copy(self):
-        return CircularBend(self.position, self.inner_radius, self.outer_radius, 
-                            self.angle, self.rotation, self.material, self.color, self.optimize, 
-                            self.points, self.depth, self.z)
-
-class Taper(Polygon):
-    """Taper is a structure that tapers from a width to a height."""
-    def __init__(self, position=(0,0), input_width=1, output_width=0.5, length=1, material=None, color=None, optimize=False, depth=0, z=0):
-        # Handle 2D position input (x,y) by adding z=0
-        if len(position) == 2: position = (position[0], position[1], 0.0)
-        elif len(position) == 3: position = position
-        else: raise ValueError("Position must be (x,y) or (x,y,z)")
-        # Calculate vertices for the trapezoid shape in 3D
-        x, y, z = position
-        vertices = [(x, y - input_width/2, z),  # Bottom left
-                    (x + length, y - output_width/2, z),  # Bottom right
-                    (x + length, y + output_width/2, z),  # Top right
-                    (x, y + input_width/2, z)] # Top left
-        super().__init__(vertices=vertices, material=material, color=color, optimize=optimize, depth=depth, z=z)
-        self.position = position
-        self.input_width = input_width
-        self.output_width = output_width
-        self.length = length
-        self.optimize = optimize
-
-    def rotate(self, angle, axis='z', point=None):
-        """Rotate the taper around its center of mass or specified point."""
-        # Use parent class rotation method
-        super().rotate(angle, axis, point)
-        # Calculate new bounding box after rotation
-        min_x = min(v[0] for v in self.vertices)
-        min_y = min(v[1] for v in self.vertices)
-        min_z = min(v[2] for v in self.vertices)
-        max_x = max(v[0] for v in self.vertices)
-        max_y = max(v[1] for v in self.vertices)
-        max_z = max(v[2] for v in self.vertices)
-        # Update position to left bottom corner and update length
-        self.position = (min_x, min_y, min_z)
-        self.length = max_x - min_x
-        return self
-
-    def copy(self):
-        """Create a copy of this taper with the same attributes and vertices."""
-        new_taper = Taper(self.position, self.input_width, self.output_width, 
-                          self.length, self.material, self.color, self.optimize, self.depth, self.z)
-        # Ensure vertices are copied exactly as they are (important for rotated tapers)
-        new_taper.vertices = [(x, y, z) for x, y, z in self.vertices]
-        return new_taper
-
-class PML:
-    """Unified PML (Perfectly Matched Layer) class for absorbing boundary conditions."""
-    def __init__(self, region_type, position, size, orientation, polynomial_order=2.0, sigma_factor=1.0, alpha_max=0.1):
-        self.region_type = region_type  # "rect" or "corner"
-        self.position = position
-        self.orientation = orientation
-        self.polynomial_order = polynomial_order  # Reduced to allow smoother transition
-        self.sigma_factor = sigma_factor  # Reduced to allow waves to enter
-        self.alpha_max = alpha_max  # Reduced frequency-shifting for smoother transition
-        if region_type == "rect": self.width, self.height = size
-        else: self.radius = size
-
-    def add_to_plot(self, ax, facecolor='none', edgecolor="black", alpha=0.5, linestyle='--'):
-        """Delegate PML drawing to beamz.viz.draw_pml."""
-        return viz.draw_pml(ax, self, facecolor=facecolor, edgecolor=edgecolor, alpha=alpha, linestyle=linestyle)
-
-    def get_profile(self, normalized_distance):
-        """Calculate PML absorption profile using gradual grading."""
-        # Ensure distance is within [0,1]
-        d = min(max(normalized_distance, 0.0), 1.0)
-        # Create a smooth transition from 0 at the interface
-        # Start with nearly zero conductivity at the interface and gradually increase
-        if d < 0.05: sigma = 0.01 * (d/0.05)**2
-        else: sigma = ((d - 0.05) / 0.95)**self.polynomial_order
-        # Smooth frequency-shifting profile
-        alpha = self.alpha_max * d**2  # Quadratic profile for smooth transition
-        return sigma, alpha
-    
-    def get_conductivity(self, x, y, dx=None, dt=None, eps_avg=None):
-        """Calculate PML conductivity at a point using smooth-transition PML."""
-        # Calculate theoretical optimal conductivity based on impedance matching
-        if dx is not None and eps_avg is not None:
-            # Calculate impedance 
-            eta = np.sqrt(MU_0 / (EPS_0 * eps_avg))
-            # Optimal conductivity for minimal reflection at interface
-            # Reduced from 1.2 to 0.8 for smoother transition
-            sigma_max = 1.2 / (eta * dx)
-            sigma_max *= self.sigma_factor  # Apply gentler factor
-        else: sigma_max = 1.0  # Lower default conductivity
-        
-        # Get normalized distance based on region type and orientation
-        if self.region_type == "rect":
-            # Check if point is within rectangular PML region
-            if not (self.position[0] <= x <= self.position[0] + self.width and
-                    self.position[1] <= y <= self.position[1] + self.height):
-                return 0.0
-            # Calculate normalized distance from boundary based on orientation
-            # Distance should be 0 at inner boundary and 1 at outer boundary
-            if self.orientation == "left": distance = 1.0 - (x - self.position[0]) / self.width
-            elif self.orientation == "right": distance = (x - self.position[0]) / self.width
-            elif self.orientation == "top": distance = (y - self.position[1]) / self.height
-            elif self.orientation == "bottom": distance = 1.0 - (y - self.position[1]) / self.height
-            else: return 0.0
-        
-        else: # corner PML
-            # Calculate distance from corner to point
-            distance_from_corner = np.hypot(x - self.position[0], y - self.position[1])
-            # Outside the PML region
-            if distance_from_corner > self.radius: return 0.0
-            # Check if in correct quadrant
-            dx_from_corner = x - self.position[0]
-            dy_from_corner = y - self.position[1]
-            if self.orientation == "top-left" and (dx_from_corner > 0 or dy_from_corner < 0): return 0.0
-            elif self.orientation == "top-right" and (dx_from_corner < 0 or dy_from_corner < 0): return 0.0
-            elif self.orientation == "bottom-left" and (dx_from_corner > 0 or dy_from_corner > 0): return 0.0
-            elif self.orientation == "bottom-right" and (dx_from_corner < 0 or dy_from_corner > 0): return 0.0
-            # Normalize distance (0 at inner edge, 1 at corner)
-            distance = distance_from_corner / self.radius
-        
-        # Get optimized profile values
-        sigma_profile, alpha_profile = self.get_profile(distance)
-        # Apply stretched-coordinate PML with gradual absorption
-        conductivity = sigma_max * sigma_profile
-        # The material-dependent scaling might have been causing excessive reflection
-        # We'll use a gentler approach that smoothly transitions at the boundary
-        if dt is not None:
-            # Apply frequency-shifting with reduced effect near boundary
-            frequency_factor = 1.0 / (1.0 + alpha_profile)
-            conductivity *= frequency_factor
-            
-        return conductivity
-    
-    def copy(self):
-        """Create a deep copy of the PML boundary."""
-        return PML(
-            region_type=self.region_type,
-            position=self.position,
-            size=(self.width, self.height) if self.region_type == "rect" else self.radius,
-            orientation=self.orientation,
-            polynomial_order=self.polynomial_order,
-            sigma_factor=self.sigma_factor,
-            alpha_max=self.alpha_max
-        )
